@@ -20,12 +20,12 @@
   "homepage": "https://github.com/kogsio/littletiers#readme",
   "dependencies": {
     "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "faker": "^5.1.0",
+    "express": "^4.18.2",
+    "faker": "^5.5.3",
     "lowdb": "^1.0.0"
   },
   "engines": {
-    "node": "14.x"
+    "node": "18.x"
   },
   "devDependencies": {
     "@faker-js/faker": "^7.6.0"
